+++ catalog/app/containers/Bucket/Dir.tsx
@@ -21,7 +21,6 @@ import * as NamedRoutes from 'utils/NamedRoutes'
 import * as BucketPreferences from 'utils/BucketPreferences'
 import parseSearch from 'utils/parseSearch'
 import * as s3paths from 'utils/s3paths'
-import type * as workflows from 'utils/workflows'
 
 import Code from './Code'
 import * as FileView from './FileView'
@@ -250,6 +249,14 @@ function DirContents({ response, locked, bucket, path, loadMore }: DirContentsPr
   )
 }
 
+const CREATE_PACKAGE_UI = {
+  successTitle: 'Package created',
+  successRenderMessage: ({ packageLink }: { packageLink: React.ReactNode }) => (
+    <>Package {packageLink} successfully created</>
+  ),
+  title: 'Create package from directory',
+}
+
 const useStyles = M.makeStyles((t) => ({
   crumbs: {
     ...t.typography.body1,
@@ -275,20 +282,14 @@ const useStyles = M.makeStyles((t) => ({
 
 interface DirParams {
   bucket: string
-  path?: string
+  path: string
+  prefix?: string
 }
 
-export default function Dir({
-  match: {
-    params: { bucket, path: encodedPath = '' },
-  },
-  location: l,
-}: RRDom.RouteComponentProps<DirParams>) {
+function Dir({ bucket, path, prefix }: DirParams) {
   const classes = useStyles()
   const s3 = AWS.S3.use()
   const prefs = BucketPreferences.use()
-  const { prefix } = parseSearch(l.search)
-  const path = s3paths.decode(encodedPath)
   const dest = path ? basename(path) : bucket
 
   const code = React.useMemo(
@@ -347,20 +348,13 @@ export default function Dir({
     )
   }, [data.result])
 
-  const packageDirectoryDialog = PD.usePackageCreationDialog({
-    bucket,
-    delayHashing: true,
-    disableStateDisplay: true,
-  })
-
+  const createPackage = PD.useCreatePackage()
   const openPackageCreationDialog = React.useCallback(
-    (successor: workflows.Successor) => {
-      packageDirectoryDialog.open({
-        path,
-        successor,
-      })
+    (s) => {
+      createPackage.setDst(s)
+      createPackage.open()
     },
-    [packageDirectoryDialog, path],
+    [createPackage],
   )
 
   const { urls } = NamedRoutes.use<RouteMap>()
@@ -373,14 +367,6 @@ export default function Dir({
   return (
     <M.Box pt={2} pb={4}>
       <MetaTitle>{[path || 'Files', bucket]}</MetaTitle>
-
-      {packageDirectoryDialog.render({
-        successTitle: 'Package created',
-        successRenderMessage: ({ packageLink }) => (
-          <>Package {packageLink} successfully created</>
-        ),
-        title: 'Create package from directory',
-      })}
 
       <div className={classes.topbar}>
         <div className={classes.crumbs} onCopy={BreadCrumbs.copyWithoutSpaces}>
@@ -443,5 +429,31 @@ export default function Dir({
         },
       })}
     </M.Box>
+  )
+}
+
+interface DirRouteParams {
+  bucket: string
+  path?: string
+}
+
+export default function DirWrapper({
+  match: {
+    params: { bucket, path: encodedPath = '' },
+  },
+  location: l,
+}: RRDom.RouteComponentProps<DirRouteParams>) {
+  const path = s3paths.decode(encodedPath)
+  const { prefix } = parseSearch(l.search, true)
+  return (
+    <PD.Provider
+      bucket={bucket}
+      s3Path={path}
+      ui={CREATE_PACKAGE_UI}
+      delayHashing
+      disableStateDisplay
+    >
+      <Dir bucket={bucket} path={path} prefix={prefix} />
+    </PD.Provider>
   )
 }