+++ catalog/app/containers/Bucket/Queries/Athena/CreatePackage.tsx
@@ -3,7 +3,7 @@ import * as M from '@material-ui/core'
 
 import * as Dialog from 'components/Dialog'
 import * as AddToPackage from 'containers/AddToPackage'
-import { usePackageCreationDialog } from 'containers/Bucket/PackageDialog/PackageCreationForm'
+import { useCreatePackage } from 'containers/Bucket/PackageDialog/Provider'
 import type * as Model from 'model'
 import * as s3paths from 'utils/s3paths'
 
@@ -122,26 +122,21 @@ const useStyles = M.makeStyles((t) => ({
 }))
 
 interface CreatePackageProps {
-  bucket: string
   queryResults: requests.athena.QueryResultsResponse
 }
 
-export default function CreatePackage({ bucket, queryResults }: CreatePackageProps) {
+export default function CreatePackage({ queryResults }: CreatePackageProps) {
   const classes = useStyles()
   const [entries, setEntries] = React.useState<ParsedRows>({ valid: {}, invalid: [] })
   const addToPackage = AddToPackage.use()
-  const createDialog = usePackageCreationDialog({
-    bucket,
-    delayHashing: true,
-    disableStateDisplay: true,
-  })
+  const createPackage = useCreatePackage()
   const handleConfirm = React.useCallback(
     (ok: boolean) => {
       if (!ok) return
       addToPackage?.merge(entries.valid)
-      createDialog.open()
+      createPackage.open()
     },
-    [addToPackage, entries, createDialog],
+    [addToPackage, entries, createPackage],
   )
   const confirm = Dialog.useConfirm({
     title: 'These rows will be discarded. Confirm creating package?',
@@ -157,9 +152,9 @@ export default function CreatePackage({ bucket, queryResults }: CreatePackagePro
       confirm.open()
     } else {
       addToPackage?.merge(parsed.valid)
-      createDialog.open()
+      createPackage.open()
     }
-  }, [addToPackage, confirm, createDialog, queryResults])
+  }, [addToPackage, confirm, createPackage, queryResults])
 
   if (!doQueryResultsContainManifestEntries(queryResults)) {
     return <SeeDocsForCreatingPackage />
@@ -167,13 +162,6 @@ export default function CreatePackage({ bucket, queryResults }: CreatePackagePro
 
   return (
     <>
-      {createDialog.render({
-        successTitle: 'Package created',
-        successRenderMessage: ({ packageLink }) => (
-          <>Package {packageLink} successfully created</>
-        ),
-        title: 'Create package',
-      })}
       {confirm.render(
         <Results
           className={classes.results}